+++ CHANGELOG.md
@@ -4,6 +4,10 @@ We follow [Semantic Versions](https://semver.org/).
 
 ## Unreleased
 
+- Refactor and improve validation on keys in django module
+- Rename `S3KeyWithUUID` to `WithPrefixUUIDFileName`
+- Rename `S3KeyWithPrefix` to `WithPrefixUUIDFolder`
+
 ## 0.1.0
 
 - Beta release